Replaced channel separator: used '@' instead of '.' used by VLANs Added support for channel range. Now the interface name is ethX@channel. Example eth0@0 eth0@0-2 eth0@1,2 eth0@0-3,5-7
git-svn-id: https://svn.ntop.org/svn/ntop/trunk/PF_RING@3586 a02cd0c1-8e76-42e1-a119-56f9641475e2

+++ userland/lib/pfring.h
@@ -118,7 +118,7 @@ extern "C" {
   /* ********************************* */
 
   int pfring_set_cluster(pfring *ring, u_int clusterId);
-  int pfring_set_channel_id(pfring *ring, short channel_id);
+  int pfring_set_channel_id(pfring *ring, int32_t channel_id);
   int pfring_remove_from_cluster(pfring *ring);
   int pfring_set_reflector(pfring *ring, char *reflectorDevice);
   pfring* pfring_open(char *device_name, u_int8_t promisc, u_int32_t caplen, u_int8_t reentrant);
